♻️refactor : getExpoInfo, getItems 메소드 각 모델로 이동
