+++ Expo1900/Expo1900/Model/Item.swift
@@ -5,6 +5,8 @@
 //  Created by Doogie, Red on 2022/04/13.
 //
 
+import UIKit
+
 struct Item: Decodable {
     let name: String
     let imageName: String
@@ -16,5 +18,15 @@ struct Item: Decodable {
         case imageName = "image_name"
         case shortDescription = "short_desc"
         case description = "desc"
+    }
+    
+    static func getItems(view: UIViewController) -> [Item] {
+        var items: [Item] = []
+        do {
+            items = try [Item].decode(with: "items")
+        } catch let error {
+            view.showAlert(for: "경고", message: "데이터 로드 오류 \n" + error.localizedDescription)
+        }
+        return items
     }
 }
